+++ babel.config.json
@@ -1,21 +1,15 @@
 {
-  "presets": [
+  
+  "plugins": [
     [
-      "@babel/preset-env",
+      "@babel/plugin-transform-typescript",
       {
-        "targets": {
-          "node": true
-        }
+        "allowDeclareFields": true
       }
     ],
-    "@babel/preset-react",
-    "@babel/preset-typescript"
-  ],
-  "plugins": [
     "@babel/plugin-proposal-class-properties",
     "@babel/plugin-proposal-export-default-from",
     "transform-react-remove-prop-types",
-    "@babel/plugin-transform-typescript",
     "@babel/plugin-proposal-object-rest-spread",
     "@babel/plugin-transform-parameters",
     [
@@ -49,5 +43,15 @@
     ],
     "@babel/plugin-proposal-nullish-coalescing-operator",
     "@babel/plugin-proposal-do-expressions"
+  ],
+  "presets": [
+    [
+      "@babel/preset-env",
+      {
+        "modules": false
+      }
+    ],
+    "@babel/preset-react",
+    "@babel/preset-typescript"
   ]
 }
